+++ src/main/java/io/github/redstoneparadox/oaktree/client/gui/control/PaddingControl.java
@@ -4,10 +4,10 @@ import io.github.redstoneparadox.oaktree.client.gui.ControlGui;
 import io.github.redstoneparadox.oaktree.client.gui.util.ScreenVec;
 
 public abstract class PaddingControl<C extends PaddingControl> extends Control<C> {
-    public float topPadding = 0.0f;
-    public float bottomPadding = 0.0f;
-    public float leftPadding = 0.0f;
-    public float rightPadding = 0.0f;
+    public int topPadding = 0;
+    public int bottomPadding = 0;
+    public int leftPadding = 0;
+    public int rightPadding = 0;
 
     /**
      * Sets the padding for all 4 sides.
@@ -15,7 +15,7 @@ public abstract class PaddingControl<C extends PaddingControl> extends Control<C
      * @param padding The padding value.
      * @return The control itself.
      */
-    public C padding(float padding) {
+    public C padding(int padding) {
         this.topPadding = padding;
         this.bottomPadding = padding;
         this.leftPadding = padding;
@@ -29,7 +29,7 @@ public abstract class PaddingControl<C extends PaddingControl> extends Control<C
      * @param topPadding The padding value.
      * @return The control itself.
      */
-    public C topPadding(float topPadding) {
+    public C topPadding(int topPadding) {
         this.topPadding = topPadding;
         return (C) this;
     }
@@ -40,7 +40,7 @@ public abstract class PaddingControl<C extends PaddingControl> extends Control<C
      * @param bottomPadding The padding value.
      * @return The control itself.
      */
-    public C bottomPadding(float bottomPadding) {
+    public C bottomPadding(int bottomPadding) {
         this.bottomPadding = bottomPadding;
         return (C) this;
     }
@@ -51,7 +51,7 @@ public abstract class PaddingControl<C extends PaddingControl> extends Control<C
      * @param leftPadding The padding value.
      * @return The control itself.
      */
-    public C leftPadding(float leftPadding) {
+    public C leftPadding(int leftPadding) {
         this.leftPadding = leftPadding;
         return (C) this;
     }
@@ -62,26 +62,26 @@ public abstract class PaddingControl<C extends PaddingControl> extends Control<C
      * @param rightPadding The padding value.
      * @return The control itself.
      */
-    public C rightPadding(float rightPadding) {
+    public C rightPadding(int rightPadding) {
         this.rightPadding = rightPadding;
         return (C) this;
     }
 
     @Override
-    public void preDraw(int mouseX, int mouseY, float deltaTime, ControlGui gui, float offsetX, float offsetY, float containerWidth, float containerHeight) {
+    public void preDraw(int mouseX, int mouseY, float deltaTime, ControlGui gui, int offsetX, int offsetY, int containerWidth, int containerHeight) {
         super.preDraw(mouseX, mouseY, deltaTime, gui, offsetX, offsetY, containerWidth, containerHeight);
     }
 
-    ScreenVec innerDimensions(float spaceWidth, float spaceHeight) {
-        float innerWidth = spaceWidth - leftPadding - rightPadding;
-        float innerHeight = spaceHeight - topPadding - bottomPadding;
+    ScreenVec innerDimensions(int spaceWidth, int spaceHeight) {
+        int innerWidth = spaceWidth - leftPadding - rightPadding;
+        int innerHeight = spaceHeight - topPadding - bottomPadding;
 
         return new ScreenVec(innerWidth, innerHeight);
     }
 
-    ScreenVec innerPosition(float spaceX, float spaceY) {
-        float innerX = spaceX + leftPadding;
-        float innerY = spaceY + topPadding;
+    ScreenVec innerPosition(int spaceX, int spaceY) {
+        int innerX = spaceX + leftPadding;
+        int innerY = spaceY + topPadding;
 
         return new ScreenVec(innerX, innerY);
     }
